Fix implicit pointer->int conversion type

+++ junk.c
@@ -1037,7 +1037,8 @@ clsf(struct message *m)
 	float	a = 1, b = 1, r;
 
 	if (verbose)
-		fprintf(stderr, "Examining message %d\n", m - &message[0] + 1);
+		fprintf(stderr, "Examining message %d\n",
+				(int)(m - &message[0] + 1));
 	for (i = 0; i < BEST; i++) {
 		best[i].dist = 0;
 		best[i].prob = -1;
@@ -1065,7 +1066,7 @@ clsf(struct message *m)
 	r = a+b > 0 ? a / (a+b) : 0;
 	if (verbose)
 		fprintf(stderr, "Junk probability of message %d: %g\n",
-				m - &message[0] + 1, r);
+				(int)(m - &message[0] + 1), r);
 	if (r > THR)
 		m->m_flag |= MJUNK;
 	else
